requirements.txt
asgiref==3.5.2
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
mysqlclient==2.1.1
platformdirs==2.5.2
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.3
whitenoise==6.2.0
